[ADD] Create pyper module to enhance access rights handling
This modules aims to make user permission handling easier. For now, we added the possibility to add to the "options" of a field 'group_action': 'readonly'.
Doing so will make the field readonly instead of making it disapear from the DOM.

This is useful for handling readonly rights on given fields of a form, while making it still visible for the concerned user.