[ADD] Add a custom animation to a statusbar widget #PYP121
Allows you to define specific SVG animations for each statusbar status.

OWL integration of the effect service to manage the display of animations with customisable parameters such as the SVG URL, message and fade-out options.

A detailed Dooki sheet is available on adding SVG animations to statuses.

Add static and dynamic effect